+++ generators/app/utils/locale.enum.js
@@ -1,0 +1,12 @@
+module.exports = {
+  en: {
+    name: 'English',
+    value: 'en',
+    checked: true,
+  },
+  fr: {
+    name: 'French',
+    value: 'fr',
+    checked: false,
+  },
+};